fixed form creator display to make it more usable during horizontal layout.
added some titles to the display
some more icons to the formreviewer

but main point of this checkin is the daterange picker activity.
we can now do a date range with two sliders that lets it be a bit easier to pick date ranges for messages and charts.

right now activity returns the results and is startable, but its full integration into charts and the dashboard needs to be fleshed out